# src/main/java/com/maheer9272/LedgerCore/controller/UserController.java
package com.maheer9272.LedgerCore.controller;

import com.maheer9272.LedgerCore.dto.UserProfileResponse;
import com.maheer9272.LedgerCore.dto.UserProfileResponseUsingAccountNumber;
import com.maheer9272.LedgerCore.dto.UserUpdateRequestDto;
import com.maheer9272.LedgerCore.dto.UserUpdateResponseDto;
import com.maheer9272.LedgerCore.service.UserService;
import jakarta.validation.Valid;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.security.core.Authentication;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping("/api/v1/users")
public class UserController {

    private final UserService userService;

    public UserController(UserService userService) {
        this.userService = userService;
    }

    @GetMapping("/me/{accountNumber}")
    public ResponseEntity<UserProfileResponseUsingAccountNumber> profileAccountNumber(
            @PathVariable String accountNumber,
            Authentication authentication) {
        UserProfileResponseUsingAccountNumber response = userService.getProfileUsingAccountNumber(accountNumber, authentication);
        return ResponseEntity
                .status(HttpStatus.OK)
                .body(response);
    }
    @GetMapping("/me")
    public ResponseEntity<UserProfileResponse> profile(
            Authentication authentication) {
        UserProfileResponse response = userService.getProfile(authentication);
        return ResponseEntity
                .status(HttpStatus.OK)
                .body(response);
    }

    @PatchMapping("/me")
    public ResponseEntity<UserUpdateResponseDto> updateUser(
            @Valid @RequestBody UserUpdateRequestDto userUpdateRequestDto,
            Authentication authentication) {

        UserUpdateResponseDto responseDto = userService.updateUser(userUpdateRequestDto, authentication);

        return ResponseEntity
                .status(HttpStatus.OK)
                .body(responseDto);
    }
}