# src/main/java/com/maheer9272/LedgerCore/dto/UserProfileResponseUsingAccountNumber.java
package com.maheer9272.LedgerCore.dto;

import lombok.AllArgsConstructor;
import lombok.Getter;

import java.math.BigDecimal;

@Getter
@AllArgsConstructor
public class UserProfileResponseUsingAccountNumber {
    private String name;
    private String email;
    private BigDecimal balance;
}