# src/main/resources/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Financial Transaction Ledger API",
    "description": "REST APIs for a fintech backend implementing user onboarding, account management, double-entry ledger transactions (deposit, withdrawal, transfer), transaction history, and balance reconciliation. All money-movement endpoints require an Idempotency-Key header to safely support retries.",
    "version": "1.0.3",
    "contact": {
      "name": "Shaik Maheer",
      "url": "https://github.com/Maheer9272/financial-transaction-ledger"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/Maheer9272/financial-transaction-ledger/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "tags": ["User"],
        "operationId": "profile",
        "summary": "Get the authenticated user's profile",
        "description": "Returns the currently authenticated user's basic profile details (not tied to a specific account).",
        "responses": {
          "200": {
            "description": "Profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserProfileResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "patch": {
        "tags": ["User"],
        "operationId": "updateUser",
        "summary": "Update the authenticated user's profile",
        "description": "Partially updates the name and/or email of the currently authenticated user.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserUpdateRequestDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserUpdateResponseDto" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/api/v1/users/me/{accountNumber}": {
      "get": {
        "tags": ["User"],
        "operationId": "profileAccountNumber",
        "summary": "Get user profile with a specific account's balance",
        "description": "Returns the authenticated user's profile details along with the balance of the given account number.",
        "parameters": [
          { "$ref": "#/components/parameters/AccountNumberPath" }
        ],
        "responses": {
          "200": {
            "description": "Profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserProfileResponseUsingAccountNumber" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from /api/v1/auth/login, sent as 'Authorization: Bearer <token>'."
      }
    },
    "parameters": {
      "AccountNumberPath": {
        "name": "accountNumber",
        "in": "path",
        "required": true,
        "description": "Unique account number",
        "schema": { "type": "string", "maxLength": 8 }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "The request failed validation",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponseDto" }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid authentication credentials",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponseDto" }
          }
        }
      },
      "Forbidden": {
        "description": "The authenticated user does not have access to this resource",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponseDto" }
          }
        }
      },
      "NotFound": {
        "description": "The requested resource could not be found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponseDto" }
          }
        }
      },
      "Conflict": {
        "description": "The request conflicts with the current state of the resource (e.g. duplicate email, reused idempotency key with a different payload)",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponseDto" }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "timestamp": { "type": "string", "format": "date-time" },
          "status": { "type": "integer", "format": "int32" },
          "error": { "type": "string" },
          "message": { "type": "string" },
          "path": { "type": "string" }
        }
      },
...
      "UserUpdateRequestDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "maxLength": 50, "minLength": 3 },
          "email": { "type": "string", "format": "email", "maxLength": 50, "minLength": 0 }
        }
      },
      "UserUpdateResponseDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string" }
        }
      },
      "UserProfileResponse": {
        "type": "object",
        "description": "Basic authenticated-user profile, not scoped to any single account.",
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string" }
        }
      },
      "UserProfileResponseUsingAccountNumber": {
        "type": "object",
        "description": "User profile combined with the balance of a specific account.",
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string" },
          "balance": { "type": "number" }
        }
      },
...
    }
  },
  "security": [
    { "bearerAuth": [] }
  ]
}